Decompose sdk-shared (#637)
* @applitools/scripts@1.0.0

* @applitools/test-server@1.0.1

* @applitools/sdk-shared@0.9.0

* @applitools/sdk-shared@0.9.1

* @applitools/sdk-shared@0.9.2

* @applitools/utils@1.2.0

* @applitools/test-server@1.0.2

+++ packages/eyes-webdriverio-5/test/it/spec-driver.spec.js
@@ -18,7 +18,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('isDriver(driver)', async () => {
@@ -140,7 +140,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('getWindowRect()', async () => {
@@ -176,7 +176,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('getWindowRect()', async () => {
@@ -218,7 +218,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('isMobile()', async () => {
@@ -255,7 +255,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('isMobile()', async () => {
@@ -289,7 +289,7 @@ describe('spec driver', async () => {
     })
 
     after(async () => {
-      await destroyBrowser()
+      if (destroyBrowser) await destroyBrowser()
     })
 
     it('isDriver(driver)', async () => {
